fix: firefox. Fix opening links in firefox.
Do not use '--new-instance' option in desktop files. When firefox has
been started to open a link, it shouldn't use '--new-instance' option,
because in that case it can't open it in already running firefox
instance.

Recent firefox correctly starts separate instances, when profile
explicitly specified, so there is no need for this option.