fix: console. Convert all default boolean values to bool explicitly.
When i specify default boolean variable on cmd with '-e', it's type will
be string and value will not be converted to bool automatically. I.e.
'false' and 'true' will be both true, because string is not empty. Now i
convert all default boolean variables to bool explicitly and use
internal values everywhere.